espfix: Sprinkle paravirt magic dust.
We are using 'populate_pgd' and 'paravirt_alloc_[pud|pmd|pte]'
functions to help the underlaying platform (such as Xen) to setup
page tables. Details at the bottom.

We could have also used 'populate_[pud|pmd|pte]' instead of
'paravirt_alloc_[pud|pmd|pte]', but upon closer inspection the
'populate' variants are used to set up one entry. While we want
to setup multiple entries. As such we could also define
a new macro - populate_[puds|pmds|ptes]' which would do the job
but I think with the 'paravirt_' functions it is easier to
spot the magic dust.

Details: Xen PV ABI has a requirement that anytime any pagetable is
added in an live pagetable it has to be read only. This is
needed so that the Xen can verify its content. As such before
we hook PMDs, PUDs, and PTEs we need to do the right dance steps.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>